Feat:뉴스레터 마지막 메일 조회 API 추가, 카테고리별 뉴스레터 조회 API 추가, 슬랙연동시 최근3개의 메일소식을 전달로직 추가

+++ backend/user/repository.py
@@ -1,5 +1,9 @@
 from backend.common.database.connector import MysqlCRUDTemplate
 from backend.common.database.model import SubscribeModel, UserModel
+from backend.common.exceptions import (
+    AlreadySubscribedException,
+    NotSubscribedNewsletterException,
+)
 from backend.user.domain import User
 
 
@@ -68,6 +72,13 @@ class UserRepository:
             super().__init__()
 
         def execute(self):
+            if (
+                self.session.query(SubscribeModel)
+                .filter(SubscribeModel.user_id == self.user.id)
+                .filter(SubscribeModel.newsletter_id == self.newsletter_id)
+                .first()
+            ):
+                raise AlreadySubscribedException
             subscribe_model = SubscribeModel(
                 id=None, newsletter_id=self.newsletter_id, user_id=self.user.id
             )
@@ -86,6 +97,8 @@ class UserRepository:
                 .filter(SubscribeModel.user_id == self.user.id)
                 .filter(SubscribeModel.newsletter_id == self.newsletter_id)
             ).first()
+            if not subscribe_model:
+                raise NotSubscribedNewsletterException
             self.session.delete(subscribe_model)
             self.session.commit()
